backend/requirements.txt
fastapi
uvicorn
python-socketio[asyncio_server]==5.7.2
vosk
ctranslate2
transformers